- Almacenamiento en Room de categorias - Almacenamiento en Room de peliculas - Almacenamiento en Room de imagenes - Almacenamiento en Room de palabras clave - Almacenamiento en Room de relacion peliculas-categorias - Almacenamiento en Room de relacion peliculas-imagenes - Almacenamiento en Room de peliculas-palabras clave

+++ app/src/main/java/com/jgt/wizelinebaz2023/storage/local/room/entities/base/MoviesTable.kt
@@ -15,7 +15,7 @@ data class MoviesTable(
     val budget:           Int,
     val homepage:         String? = null,
     val overview:         String? = null,
-    val popularity:       Int,
+    val popularity:       Float,
     val revenue:          Int,
     val runtime:          Int? = null,
     val status:           String,
@@ -27,6 +27,6 @@ data class MoviesTable(
     val originalTitle:    String,
     val posterPath:       String? = null,
     val releaseDate:      String,
-    val voteAverage:      Int,
+    val voteAverage:      Float,
     val voteCount:        Int,
 )
